feat: Enhance item management and reporting features
- Refactor createItemService to improve item creation and updating logic.
- Add getItemLocationServices to retrieve item location details.
- Implement deleteLocationService to prevent deletion of locations with assigned items.
- Update stock movement services to consolidate stock adjustment logic.
- Introduce new ItemDetailsPage and ItemReportPage for detailed item views and reporting.
- Add ItemReportFilter component for selecting items and generating reports.
- Enhance StockHistoryTable to display item details and stock movements.
- Update routing and navigation to support new item detail and report pages.
- Improve UI components for better user experience in item and stock management.

+++ backend/services/location.services.js
@@ -90,6 +90,10 @@ export const updateLocationStatusService = async (locationId, status) => {
 };
 
 export const deleteLocationService = async (locationId) => {
+  const itemExist = await Item.exists({"inventory.locationId" : locationId})
+  if(itemExist){
+    throw new ApiError(HTTP_STATUS.BAD_REQUEST, "Cannot delete location: Items are still assigned to this location");
+  }
   const location = await Location.findByIdAndDelete(locationId);
   if (!location) {
     throw new ApiError(HTTP_STATUS.NOT_FOUND, "Location not found");
